Updating the Asset Management System.
Major updates involve how textures are stored. I made it so all textures
are stored in a Texture class that contains the texture and the type of
texture.

Other than that, I updated the Asset Manager to store all textures inside
of a single unordered_map.

+++ BosonEngine/BosonProject/GraphicsCore.cpp
@@ -250,7 +250,7 @@ void GraphicsCore::loadMeshes(std::wstring folderPath)
 		::FindClose(hFind);
 	}
 
-	for (int i = 0; i < filePaths.size(); i++)
+	for (size_t i = 0; i < filePaths.size(); i++)
 	{
 		std::wstring name = filePaths[i];
 		std::string nameTemp(name.begin(), name.end());
